Notebooks/Rest_framework/08模式和客户端库/requirement.txt
coreapi
coreapi-cli
djangorestframework
Django
